+++ packages/components/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @equinor/mad-components
 
+## 0.6.3
+
+### Patch Changes
+
+-   aba4173: fixed onchange and onblur issue
+
+## 0.6.2
+
+### Patch Changes
+
+-   1499432: fixed input onchange issue
+
 ## 0.6.1
 
 ### Patch Changes