+++ packages/navigation/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @equinor/mad-navigation
 
+## 0.1.8
+
+### Patch Changes
+
+-   Updated dependencies [aba4173]
+    -   @equinor/mad-components@0.6.3
+
+## 0.1.7
+
+### Patch Changes
+
+-   Updated dependencies [1499432]
+    -   @equinor/mad-components@0.6.2
+
 ## 0.1.6
 
 ### Patch Changes